organized files for job controller get all search query functions, functioing
