+++ queries/jobs.js
@@ -1,75 +1,23 @@
 const db = require("../db/dbConfig.js");
 const { createJobSkill, deleteAllJobSkills } = require("./jobSkills.js");
 
+
 const getAllJobs = async (
   limitValue,
   startValue,
+  dbQueryString,
   input,
   city,
   remote,
-  skills,
-  skillCount
+  skillsObj,
+ 
 ) => {
-  const inputQuery = input
-    ? `  (
-    (
-  (LOWER(regexp_replace(title, ' ', '', 'g')) LIKE $3)  OR
-  (LOWER(regexp_replace(city, ' ', '', 'g')) LIKE $3)  OR
-  (LOWER(regexp_replace(details, ' ', '', 'g')) LIKE $3)  OR
-  (LOWER(regexp_replace(company, ' ', '', 'g')) LIKE $3)
-  ) 
+ const { skillDbSyntax, skillCount} = skillsObj
 
-)`
-    : null;
-
-  const cityQuery = city
-    ? `LOWER(regexp_replace(city, ' ', '', 'g')) LIKE $4`
-    : null;
-
-  const remoteQuery =
-    remote !== undefined
-      ? `
-  full_remote IS $5`
-      : null;
-
-  const whereKeyword =
-    inputQuery || cityQuery || remoteQuery || skills
-      ? `WHERE (
-  ${skills ? skills : ""}
-  ${skills && (inputQuery || cityQuery || remoteQuery) ? "AND" : ""}
-  ${inputQuery ? inputQuery : ""}
-  ${inputQuery && cityQuery ? "AND" : ""} 
-  ${cityQuery ? cityQuery : ""}
-  ${remoteQuery && (cityQuery || inputQuery) ? "AND" : ""}
-  ${remoteQuery ? remoteQuery : ""}
-)`
-      : "";
-
-  let dbCommand = skills
-    ? `SELECT 
-id, title, company, city, details, full_remote, tasks, recruiter_id
-FROM jobs
-INNER JOIN jobs_skills
-ON jobs_skills.job_id=jobs.id 
-${whereKeyword && whereKeyword}
-GROUP BY jobs.id 
-HAVING 
-COUNT(job_id) = $7
-ORDER BY id 
-LIMIT $1  
-OFFSET $2
-`
-    : `SELECT * 
-FROM jobs
-${whereKeyword && whereKeyword}
-ORDER BY id 
-LIMIT $1  
-OFFSET $2
-`;
   try {
     const allJobs = await db.any(
       `
-      ${dbCommand}
+      ${dbQueryString}
        `,
       [
         limitValue,
@@ -77,7 +25,7 @@ OFFSET $2
         `%${input}%`,
         `%${city}%`,
         remote,
-        skills,
+        skillDbSyntax,
         skillCount,
       ]
     );
@@ -87,6 +35,100 @@ OFFSET $2
     return error;
   }
 };
+
+
+
+
+
+
+
+
+// const getAllJobs = async (
+//   limitValue,
+//   startValue,
+//   input,
+//   city,
+//   remote,
+//   skills,
+//   skillCount
+// ) => {
+//   const inputQuery = input
+//     ? `  (
+//     (
+//   (LOWER(regexp_replace(title, ' ', '', 'g')) LIKE $3)  OR
+//   (LOWER(regexp_replace(city, ' ', '', 'g')) LIKE $3)  OR
+//   (LOWER(regexp_replace(details, ' ', '', 'g')) LIKE $3)  OR
+//   (LOWER(regexp_replace(company, ' ', '', 'g')) LIKE $3)
+//   ) 
+
+// )`
+//     : null;
+
+//   const cityQuery = city
+//     ? `LOWER(regexp_replace(city, ' ', '', 'g')) LIKE $4`
+//     : null;
+
+//   const remoteQuery =
+//     remote !== undefined
+//       ? `
+//   full_remote IS $5`
+//       : null;
+
+//   const whereKeyword =
+//     inputQuery || cityQuery || remoteQuery || skills
+//       ? `WHERE (
+//   ${skills ? skills : ""}
+//   ${skills && (inputQuery || cityQuery || remoteQuery) ? "AND" : ""}
+//   ${inputQuery ? inputQuery : ""}
+//   ${inputQuery && cityQuery ? "AND" : ""} 
+//   ${cityQuery ? cityQuery : ""}
+//   ${remoteQuery && (cityQuery || inputQuery) ? "AND" : ""}
+//   ${remoteQuery ? remoteQuery : ""}
+// )`
+//       : "";
+
+//   let dbCommand = skills
+//     ? `SELECT 
+// id, title, company, city, details, full_remote, tasks, recruiter_id
+// FROM jobs
+// INNER JOIN jobs_skills
+// ON jobs_skills.job_id=jobs.id 
+// ${whereKeyword && whereKeyword}
+// GROUP BY jobs.id 
+// HAVING 
+// COUNT(job_id) = $7
+// ORDER BY id 
+// LIMIT $1  
+// OFFSET $2
+// `
+//     : `SELECT * 
+// FROM jobs
+// ${whereKeyword && whereKeyword}
+// ORDER BY id 
+// LIMIT $1  
+// OFFSET $2
+// `;
+//   try {
+//     const allJobs = await db.any(
+//       `
+//       ${dbCommand}
+//        `,
+//       [
+//         limitValue,
+//         startValue,
+//         `%${input}%`,
+//         `%${city}%`,
+//         remote,
+//         skills,
+//         skillCount,
+//       ]
+//     );
+
+//     return allJobs;
+//   } catch (error) {
+//     return error;
+//   }
+// };
 
 const getSkillsForJobByJobId = async (jobId) => {
   const allSkills = await db.any(
